Move PycorsHomeProviderTrait integration tests mocking to top-module for sharing
