+++ tests/integration/mod.rs
@@ -2,7 +2,11 @@
 // executable name (integration_tests) due to a bug in rustfmt.
 // See: https://github.com/rust-lang/rustfmt/issues/3794
 
-use std::{env, fs, io::Write, path::Path};
+use std::{
+    env, fs,
+    io::Write,
+    path::{Path, PathBuf},
+};
 
 use assert_cmd::{assert::OutputAssertExt, Command};
 use indoc::indoc;
@@ -14,6 +18,7 @@ use pycors::{
         home_overwrite_env_variable, project_home_env_variable, EXECUTABLE_NAME, INFO_FILE,
         TOOLCHAIN_FILE,
     },
+    utils::directory::{PycorsHomeProviderTrait, PycorsPathsProvider},
     Result,
 };
 
@@ -27,6 +32,15 @@ mod run;
 mod select;
 mod setup;
 mod version;
+
+mockall::mock! {
+    PycorsHomeProviderTrait {}     // Name of the mock struct, less the "Mock" prefix
+    trait PycorsHomeProviderTrait {   // definition of the trait to mock
+        fn home(&self) -> Option<PathBuf>;
+        fn project_home(&self) -> Option<PathBuf>;
+        fn paths(&self) -> Vec<PathBuf>;
+    }
+}
 
 fn select(version: &str, cwd: &Path) {
     let _ = fs::create_dir_all(&cwd);
